RAG-Assist/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.10.1
langchain==0.3.13
langchain-core>=0.3.26,<0.4.0
langchain-text-splitters>=0.3.3,<0.4.0
sentence-transformers==3.3.1
pymupdf==1.25.2
pinecone-client==5.0.1
transformers==4.47.1
torch>=2.6.0
google-genai==1.3.0
google-generativeai==0.8.3
groq==1.0.0
fastapi
fastapi
uvicorn
langchain
langchain-community
pdf2image
pytesseract
unstructured[all-docs]
python-multipart
streamlit==1.41.0
requests==2.32.3
python-docx==1.1.2